requirements.txt
Flask
azure-storage-blob
azure-identity
celery[redis]
google-generativeai
openai
python-dotenv
imageio[ffmpeg]
imageio[pyav]
numpy
